can diff 2 marked files

+++ contextDiff.py
@@ -27,18 +27,22 @@ if sys.argv[1]=="1":
    if notification:
 	   notification.notify("Noted for diff", "Select 2nd file with 'diff2' to diff with "+
 		os.linesep+sys.argv[2],timeout=5,app_icon="gnome-session-switch") 
-def doit():
+
+def main():
    if sys.argv[1]=="2":
+      fn1 = None
+      if len(sys.argv) >3:
+          os.popen("meld "+sys.argv[2]+" "+sys.argv[3])
+          return
       if not os.path.exists(tmpfile):
          return
       fn2 = sys.argv[2]
-      fn1 = None
       with open(tmpfile,"r") as f:
        fn1 = f.read()
       os.popen("rm "+ tmpfile)
       if fn1:
        os.popen("meld "+fn1+" "+fn2)
 
-doit()
+main()
 
 #eof
